requirements.txt
kafka-python==2.0.2
mysql-connector-python==8.3.0
mysqlclient==2.2.4
pyee==11.1.0
python-ffmpeg==2.0.12
typing-extensions==4.11.0
flask==2.2.5
python-dotenv==1.0.1
uuid==1.30
Werkzeug==2.2.3
thread6==0.2.0
Flask-Cors==1.10.3